adding a few packets. like spectator, channel etc.

+++ packets/login_reply.go
@@ -8,6 +8,8 @@ import (
 func (pw *PacketWriter) LoginReply(reply consts.LoginReply) {
 	p := new(Packet)
 	p.PacketId = consts.ServerLoginReply
-	p.WriteData(helpers.GetBytes(reply))
+
+	p.WriteData(helpers.GetBytes(int32(reply)))
+
 	pw.WritePacket(p)
 }
